server/api/views.py
from django.http import JsonResponse
from rest_framework import status,viewsets
from django.shortcuts import render
from rest_framework.decorators import api_view,authentication_classes,permission_classes
from rest_framework.authentication import TokenAuthentication
from rest_framework.permissions import IsAuthenticated,AllowAny
from rest_framework.response import Response
from django.contrib.auth.models import User
from rest_framework.views import APIView
from rest_framework.authtoken.models import Token
from django.views.decorators.csrf import csrf_exempt
from .models import *
from .serializers import *
# Create your views here.

class ImageUploadView(APIView):
    def post(self, request, *args, **kwargs):
        image_serializer = UploadedImageSerializer(data=request.data)
        print(request.data)
        if image_serializer.is_valid():
            image_serializer.save()
            return Response(image_serializer.data, status=status.HTTP_201_CREATED)
        else:
            return Response(image_serializer.errors, status=status.HTTP_400_BAD_REQUEST)