server/requirements.txt
asgiref==3.8.1
backend==0.2.4.1
certifi==2024.6.2
Django==4.1.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.6.1
gunicorn==22.0.0
packaging==24.0
pillow==10.3.0
pymongo==3.12.3
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.2.4
